requirements.txt
# Django and related packages
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0

# Production Server
gunicorn>=21.2,<22.0
whitenoise>=6.5,<7.0

# Environment variables
python-dotenv>=1.0,<2.0

# API Clients
apify-client>=1.0,<2.0
gspread>=5.10,<6.0
google-auth-oauthlib>=1.0,<2.0
google-auth-httplib2>=0.1,<1.0

# Data processing and HTTP
pandas>=2.0,<3.0
requests>=2.31,<3.0